refactor: use api routes for infinite scroll fetching
